+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
 * Added RateMetricAgg_ for ElasticSearch
 * added axisSoftMin and axisSoftMax options for TimeSeries
 * Added support for Azure Data Explorer datasource plugin (https://github.com/grafana/azure-data-explorer-datasource)
+* Added ``sortBy`` parameter to Table panel
 
 .. _`Bar_Chart`: https://grafana.com/docs/grafana/latest/panels-visualizations/visualizations/bar-chart/
 .. _`RateMetricAgg`: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-rate-aggregation.html